pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "futures-bot"
version = "0.1.0"
description = "Deterministic futures signal bot scaffold"
readme = "README.md"
requires-python = "==3.11.*"
authors = [{ name = "Futures Bot Team" }]
dependencies = [
  "numpy",
  "pandas",
  "pyyaml",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

[project.scripts]
futures-bot = "futures_bot.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra --strict-markers"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
show_error_codes = true
pretty = true
mypy_path = "src"
packages = ["futures_bot"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
